requirements.txt
-f /usr/share/pip-wheels
amqp==2.1.4
Babel==2.3.4
beautifulsoup4==4.5.1
billiard==3.5.0.2
codecov==2.0.5
configparser==3.5.0
coverage==4.2
Django==1.9.5
django-common-helpers==0.9.0
django-cors-headers==1.3.1
django-cron==0.5.0
django-excel==0.0.5
django-filter==1.0.1
django-user-agents==0.3.0
djangorestframework==3.5.3
enum34==1.1.6
et-xmlfile==1.0.1
ezodf==0.3.2
flake8==3.2.1
get==0.0.0
html2text==2016.9.19
jdcal==1.3
kombu==4.0.2
logging==0.4.9.6
lxml==3.7.0
Markdown==2.6.7
mccabe==0.5.3
MySQL-python==1.2.5
nose==1.3.7
openpyxl==2.4.1
post==0.0.0
public==0.0.0
pycodestyle==2.2.0
pyexcel==0.3.3
pyexcel-io==0.2.5
pyexcel-ods3==0.2.2
pyexcel-text==0.2.4
pyexcel-webio==0.0.8
pyexcel-xls==0.2.3
pyexcel-xlsx==0.2.3
pyexcel-xlsxw==0.0.1
pyflakes==1.3.0
PyMySQL==0.7.9
python-memcached==1.58
pytz==2016.10
PyYAML==3.12
query-string==0.0.0
requests==2.12.4
setupfiles==0.0.0
six==1.10.0
tabulate==0.7.7
texttable==0.8.7
ua-parser==0.7.2
user-agents==1.0.1
vine==1.1.3
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.1.2
